+++ packages/common/src/models/Action.ts
@@ -80,4 +80,5 @@ export enum Action {
   CloseEngageTimePoll = 'closeEngageTimePoll',
   ShowEngageTimeSession = 'showEngageTimeSession',
   ShowEngageTimePoll = 'showEngageTimePoll',
+  SendEngageTimeMessage = 'sendEngageTimeMessage',
 }