+++ packages/common/src/models/Demos.ts
@@ -77,6 +77,9 @@ export interface Step extends IOpenWebsite, IImagePreview, ITerminal {
   // EngageTime
   sessionId?: string;
   pollId?: string;
+  startOnOpen?: boolean;
+  title?: string;
+  type?: EngageTimeMessageType;
 }
 
 export interface ITerminal {
@@ -107,3 +110,5 @@ export interface State {
   key: string;
   value: string;
 }
+
+export type EngageTimeMessageType = 'demo' | 'slide' | 'custom';